msm8974: build: Add support for Snapdragon LLVM Compiler for Krait CPUs
With this switch we can determine if we want to use Snapdragon sdclang
optimizations instead of Android Clang. Snapdragon LLVM is optimized for
Krait CPUs (400/600/800/801/805) that use ARM platform.

There's two requeriments for building with SDCLANG:

1.- TARGET_USE_SDCLANG := true should be set per device.
2.- You need to install Snapdragon LLVM Compiler toolchain (3.8.8) to
    prebuilts/clang/linux-x86/host/sdclang-3.8

If sdclang-3.8 is not present and TARGET_USE_SDCLANG is set to true
android will automatically use Android Clang instead

Change-Id: I76b34bf025072c6b69e808c21f28129d2660de87